refactor ode solvers
